requirements.txt
Appium-Python-Client==3.1.1
selenium==4.16.0
xmlrunner==1.7.7
